requirements.txt
streamlit
qrcode
pillow
python-dotenv
cloudinary